# common/api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: E-Commerce Microservices Platform API
  description: Public API surface exposed through the API Gateway (localhost:8080)
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /api/products:
    get:
      summary: List products
      tags: [Products]
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200": { description: OK }
    post:
      summary: Create a product (admin)
      tags: [Products]
      responses:
        "201": { description: Created }
  /api/products/{id}:
    get:
      summary: Get product by id
      tags: [Products]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "404": { description: Not found }
...
  /api/orders/{id}:
    get:
      summary: Get order status/details
      tags: [Orders]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
  /api/orders/{id}/cancel:
    post:
      summary: Cancel an order
      tags: [Orders]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "202": { description: Accepted }
...